Now the command buffers are filled in ctor, instead of being filled in the render loop.

+++ source/Renderer/VertexBuffer.inl
@@ -11,7 +11,7 @@ namespace renderer
 		, MemoryPropertyFlags flags )
 		: VertexBufferBase{ device
 			, bindingSlot
-			, count * sizeof( T )
+			, uint32_t( count * sizeof( T ) )
 			, target
 			, flags }
 	{
